requirements.txt
discord.py~=2.4.0
asyncpg~=0.29.0
pyyaml~=6.0.2
